UNOMI-179 Event Persistence Non-transient
Changed the persistence field to be non-transient in the Event class.
Used ESEventMixIn to prevent that value from being saved in ES.  Updated
the Event constructor to accept persistent.

This feature allows the user to pass "persistent": false with their
events if the intent is only to trigger a rule / action downstream.

If "persistent" is not set it defaults to true.